client: enable service_tokens using initialized or per-call tokens
You can create a Client instance with "service_tokens: true" parameter,
and that shall be used throughout the life of the instance to allow you
to specify "service_token: 'token'" per call. Note that this also will
require you to specify a "service_id" argument per call.

You can still use "provider_key: 'key'" when instantiating the client,
but this is now deprecated in favour of service tokens.